fix: fix #38 by exporting unstyled components
