+++ .gitignore
@@ -9,6 +9,7 @@ package-lock.json
 .cache
 .rdoc-dist
 .vscode
+.idea
 
 *.bak
 *.tem
